Add files via upload
In this task you will create a Kubernetes cluster and deploy the container to it. Kubernetes is a computing platform for managing container-based applications. They can be considered as a platform for microservices as they can manage and automate the deployment and scaling of our containers. It also known as a container orchestration platform which means the automation of much of the operational effort required to run containerized microservices. In this task, you will deploy a containerized Node.js application which you have already done in task 5.1P, onto a Kubernetes cluster and interact with the deployed application using Kubernetes commands.

The required tools for doing this task are as follows:


• Git (https://github.com)
• Visual Studio code (https://code.visualstudio.com/)
• Node.js (https://nodejs.org/en/download/)
• Docker
• Kubernetes // a computing platform to host your microservice
• Kubernetes CLI (kubectl) //You will use the Kubernetes command-line interface (kubectl) to deploy and manage your application on the Kubernetes cluster.
• Docker CLI// You will use the Docker command-line interface (CLI) to build, tag, and push the Docker image for your application to a Docker registry